created alarm object
now clocks have a list of alarms. The function alarmSound now its called checkAlarms, and it returns an int instead of a boolean. Plans of creating some booleans for the Alarm object to determine wich attributes of the datetime object are meant to take in account (ms are never taken in account).