alarms now sound
added a default alarm ringtone, and all future media should be placed in the media file. Alarm's booleans still not added and need to apply some thread management in order to make the alarm sound untill the user stops it and while that, still print the hour (this last part may be optional).